Add `not` assertion overload that takes a block
The current patterns that can be applied with the existing methods are:

```kotlin
expectThat(thing).not().and {
  isNull()
  someAssertion()
}
```

The block works as expected, but a possibly unintended side effect is if any more assertions are chained after the `and {}` block they are also inverted from the `not()`.

Another way to do this is use the `not()` builder from inside the `and {}` block:

```kotlin
expectThat(thing).and {
  not().isNull()
  not().someAssertion()
}
```

This works as expected but is fairly verbose.
This method is useful, and also works well when mixing in other assertions.

With this new `not {}` API the `not {}` builder block combines aspects of `not()` and `and {}` while retaining the asseriton strategy after the block.

```kotlin
expectThat(thing).not {
  isNull()
  someAssertion()
}
```

issue https://github.com/robfletcher/strikt/issues/139